requirements.txt
argostranslate>=1.9.0
python-docx>=0.8.11
pdfplumber>=0.9.0
Pillow>=9.0.0
